Infrastructure to have document-related functions synchronous and others asyncrhonous. See: #141
This is almost ready, but we need to check the LibspecManager.
In particular, as there is now concurrence among the processes, initializing
the APIs to generate the libspec files can timeout during tests (probably
due to running in slow machines, so, we may need to revise the timeouts
or pre-generate the builtin libraries).