Use better heuristics to create a proper test suite (considering __init__.robot in parents). Fixes #482
Note: we also make it possible for the user to specify the suiteTarget directly and use the targets
only to provide filtering.

So, when launching, if makeSuite is true, heuristics will be done based on the target to
create the suite needed with additional filtering to load the required __init__.robot
files while still running only the specified tests.